fix: comprehensive CI/CD and E2E test improvements
Major changes:

**CI/CD Workflows:**
- Add minimal E2E workflow for fast PR checks (~15-20min vs 60min+)
- Split full E2E into parallel matrix jobs (5 test suites running concurrently)
- Fix gofmt check to pass when all files are formatted (grep || true)
- Auto-format generated deepcopy files in Makefile
- Remove Prometheus from minimal E2E (not needed by selected tests)

**E2E Test Suite:**
- Add PREBUILT_WORKFLOWS env var to selectively build workflow images
  - Minimal E2E builds only 2/3 images (saves ~2-3 min)
  - Full E2E builds all 3 (default, backward compatible)
- Fix gitops PostgreSQL test: add persistence.postgresql.jdbcUrl configuration
- Fix workflow ID mismatch in test verification (use -persistence suffix)
- Reduce test timeouts from 10min to 3min where appropriate
- Fix nil pointer when workflows are selectively built

**Build & Test Fixes:**
- Fix FlowCRC type from uint32 to int32 for CRD compatibility
- Fix flaky TestSonataFlowBuildController test
- Fix Go 1.26 covdata tool error in test target
- Fix KIND image loading (use load-docker-image target)
- Free up disk space before image builds
- Update image registry to kubesmarts
- Install CEKit system dependencies in GHA

**Makefile Improvements:**
- Add deploy-only target to skip manifest regeneration
- Auto-format generated files after controller-gen
- Fix kustomize installation with correct version naming
- Add .EXPORT_ALL_VARIABLES for proper PATH handling

**Documentation:**
- Add CLI_MIGRATION_PLAN.md
- Enhanced E2E failure diagnostics with workflow pod logs

+++ utils/crc.go
@@ -23,11 +23,12 @@ import (
 	"hash/crc32"
 )
 
-func Crc32Checksum(v interface{}) (uint32, error) {
+func Crc32Checksum(v interface{}) (int32, error) {
 	var buf bytes.Buffer
 	enc := gob.NewEncoder(&buf)
 	if err := enc.Encode(v); err != nil {
 		return 0, err
 	}
-	return crc32.ChecksumIEEE(buf.Bytes()), nil
+	// Cast uint32 to int32 for Kubernetes CRD compatibility
+	return int32(crc32.ChecksumIEEE(buf.Bytes())), nil
 }
